Doxygen
sentence and document were not commented because I don't know the status
of the movie class - does it replace it, or is it in addition to it?
only source files contained more detail doxygen comments, headers
contained just the @file and @authour and short description, for
clarity.
indexer and indexing.cpp have not been commented, not sure whats going
on there right now.
abstract_tokenizer and index_item can't be abstract without breaking
everything else along the way, unfortuantely